idmap: Do not unlink the file if it is locked by another process.
The idmap file will be locked if it is used by one process. Then if another
process need to access this file, it will get the file locked state.

So in this case, should not unlink the file. Otherwise this file will be
deleted by this process, and the another process which locked it will fail
to write the file.

CRs-fixed: 763809

Change-Id: I2bca8b04dd82b5a56d30b103325ba7e22f0072d9