base: ConnectivityServer: Don't reap the new network
When we add a network it may show up as unneeded while we are
adding it and reaped as we are adding it.

Reaping it immediately can happen before the captive portal
check has completed.  This can then cause the network to get added
again causing an unnecessary battery drain.

Change-Id: I837f627c9f6d95bb8a021d2a58f5fc28564ef8b4
Signed-off-by: AudioGod <audiogod@sungsonic.co>