BatteryService: Prevent UUC scenarios at 0%
If the battery is at 0% and we're plugged to a power source, we were
assuming the device is charging and leaving it fully active. This could
lead to an unusable capacity scenario if the device was busy with high-power
tasks (camcorder, GPS, games, depends on hardware) that caused the discharge
rate to be higher than the charge rate, taking the battery all the way down
to an undervoltage lockout.

Triggering an UVLO was, in some cases, as easy as turning on GPS navigation
while using a car charger...

Change-Id: I0f805880309bf12142203badd1e33db6e37da9c1
Signed-off-by: AudioGod <audiogod@sungsonic.co>